% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmb.coordinates.R
\name{gmb.coordinates}
\alias{gmb.coordinates}
\title{For chromosome and position, calculates a genomic coordinate}
\usage{
gmb.coordinates(chr, pos, chrlen = mouse.chrlen)
}
\arguments{
\item{chr}{chromosome}

\item{pos}{position (b or Mb)}

\item{chrlen}{lengths of chromosomes as in \code{org.Mm.egCHRLENGTHS}}
}
\description{
Convert (chr, mb) coordinates to (gmb) coordinates
}
\examples{
gmb <- gmb.coordinates(Tmem68$annotation$Chr, Tmem68$annotation$Pos)
}
\seealso{
\code{\link{plot.mediation}}
}

