% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kplot.R
\name{kplot}
\alias{kplot}
\title{Karl Broman's Interactive Plot of conditioned LOD scores}
\usage{
kplot(medresults, symbol.col = "Associated.Gene.Name",
  chrlen = mouse.chrlen, ...)
}
\arguments{
\item{medresults}{mediation object}

\item{symbol.col}{column with gene names}

\item{chrlen}{length of chromosomes as in \code{org.Mm.egCHRLENGTHS}}
}
\description{
Plot LOD statistics calculated by \code{mediation.scan} against
genomic positions. Display a gene name when hovering a mouse over.
}
\examples{
data(Tmem68)
med <- mediation.scan(target=Tmem68$target,
                      mediator=Tmem68$mediator,
                      annotation=Tmem68$annotation,
                      covar=Tmem68$covar,
                      qtl.geno=Tmem68$qtl.geno)
kplot(med, symbol.col="Associated.Gene.Name")

# save the plot as HTML page
pl <- kplot(med, symbol.col="Associated.Gene.Name")
tmpdir <- tempdir()
htmlwidgets::saveWidget(pl, paste0(tmpdir, "/Tmem68.html"),
                        selfcontained=FALSE, libdir = "_assets")
}
\seealso{
\code{\link{plot.mediation}}, \code{\link{mediation.scan}}
}

