% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.scan.R
\name{mediation.scan}
\alias{mediation.scan}
\title{Mediation Scan}
\usage{
mediation.scan(target, mediator, annotation, qtl.geno, covar = NULL,
  method = c("double-lod-diff", "ignore", "lod-diff", "lod-ratio"),
  verbose = TRUE)
}
\arguments{
\item{target}{A numeric vector with gene/protein expression}

\item{mediator}{A matrix, each column is one gene/protein's expression}

\item{annotation}{A data frame with mediators' annotation, must include columns "chr" and "pos"}

\item{qtl.geno}{A matrix, haplotype probabilities at QTL we try to mediate}

\item{covar}{A matrix with additive covariates}

\item{method}{A method to handle missing cases}

\item{verbose}{If TRUE display information about the progress}
}
\description{
For a given QTL haplotype probabilities \code{qtl.geno} and target \code{target},
the function sequentially tries to add each column \code{m} of \code{mediator} matrix as a covariate
and calculates LOD statistic. The low LOD value indicates \code{qtl.geno} and
\code{target} are conditionally independent given \code{m},
i.e. \code{m} is a mediator of causal relationship from \code{qtl.geno} to \code{target}.
}
\examples{
data(Tmem68)
med <- mediation.scan(target = Tmem68$target,
                      mediator = Tmem68$mediator,
                      annotation = Tmem68$annotation,
                      covar = Tmem68$covar,
                      qtl.geno = Tmem68$qtl.geno,
                      method = "double-lod-diff")
plot(med)
}
\seealso{
\code{\link{plot.mediation}}, \code{\link{kplot}}
}
