% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_index.R
\name{mediation_index}
\alias{mediation_index}
\alias{plot.mediation_index}
\alias{autoplot.mediation_index}
\alias{ggplot_mediation_index}
\title{Test mediation across set of indexed drivers}
\usage{
mediation_index(
  target,
  mediator,
  driver = NULL,
  annotation = NULL,
  covar_tar = NULL,
  covar_med = NULL,
  driver_med = NULL,
  annotation_names = c(facet = "chr", index = "pos"),
  ...
)

\method{plot}{mediation_index}(x, ...)

\method{autoplot}{mediation_index}(x, ...)

ggplot_mediation_index(
  x,
  response = c("pvalue", "IC"),
  alpha = 0.5,
  pattern_name = "pattern",
  ...
)
}
\arguments{
\item{target}{vector or 1-column matrix with target values}

\item{mediator}{vector or 1-column matrix with mediator values}

\item{driver}{vector or matrix with driver values}

\item{annotation}{optional annotation data frame for mediators}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{driver_med}{driver array for mediators}

\item{annotation_names}{names in annotation of columns for facet, index and, optionally, driver (default `c(facet = "chr", index = "pos", driver = NULL)`)}

\item{...}{additional parameters}

\item{x}{object of class \code{mediation_index}}

\item{response}{type of response from \code{c("pvalue","IC")}}

\item{alpha}{ggplot2 transparency parameter with default \code{0.5}}

\item{pattern_name}{name of pattern with default \code{"pattern"}}
}
\description{
Test mediation across set of indexed drivers
}
\examples{
data(Tmem68)
 
target <- Tmem68$target

# Mediator is Tmem68.
m <- grep("Tmem68", Tmem68$annotation$symbol)
mediator <- Tmem68$mediator[,m]
annotation <- Tmem68$annotation[m,, drop = FALSE]

# Reduce to SNP for B6
driver_SNP <- cbind(B6 = Tmem68$driver[,2], rest = 1 - Tmem68$driver[,2])

med_index <- mediation_index(target = target,
                      mediator = mediator,
                      driver = Tmem68$driver,
                      annotation = annotation,
                      covar_tar = Tmem68$covar)
summary(med_index)
ggplot2::autoplot(med_index)

}
