% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mplot.R
\name{mplot}
\alias{mplot}
\title{Interactive Plot of conditioned LOD scores}
\usage{
mplot(medresults, symbol.col = "Associated.Gene.Name",
  chrlen = mouse.chrlen, ...)
}
\arguments{
\item{medresults}{mediation object}

\item{symbol.col}{column with gene name}

\item{chrlen}{length of chromosomes as in \code{org.Mm.egCHRLENGTHS}}
}
\description{
For a given QTL haplotype probabilities `Q` and target `T`, the function tries to add all
possible mediators `M` and calculates LOD statistic. The low LOD value indicates `Q` and
`T` are conditionally independent given `M`, i.e. M is a mediator of causal relationship
for `Q` to T.
}
\examples{
data(Tmem68)
med <- mediation.scan(target=Tmem68$target,
                      mediator=Tmem68$mediator,
                      annotation=Tmem68$annotation,
                      covar=Tmem68$covar,
                      qtl.geno=Tmem68$qtl.geno)
mplot(med, symbol.col="Associated.Gene.Name")

# save the plot as HTML page
pl <- mplot(med, symbol.col="Associated.Gene.Name")
tmpdir <- tempdir()
htmlwidgets::saveWidget(pl, paste0(tmpdir, "/Tmem68.html"),
                        selfcontained=FALSE, libdir = "_assets")
}

