% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_fun.R
\name{plot_trends}
\alias{plot_trends}
\title{Function for plotting trends in mean catch by year for species in the biomass data}
\usage{
plot_trends(biomass, species = "redfish")
}
\arguments{
\item{biomass}{Data frame with trawl catches of shrimp, cod, halibut and redfish from 1995-2010}

\item{species}{Species to select for calculations ("shrimp", "cod", "halibut" or "redfish")}
}
\value{
A plot of mean catch, +/- 95 percent confidence intervals, by year
}
\description{
Function for plotting trends in mean catch by year for species in the biomass data
}
