% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_fun.R
\name{trends}
\alias{trends}
\title{Function for calculating mean catch by year for species in the biomass data}
\usage{
trends(biomass, species = "redfish")
}
\arguments{
\item{biomass}{Data frame with trawl catches of shrimp, cod, halibut and redfish from 1995-2010}

\item{species}{Species to select for calculations ("shrimp", "cod", "halibut" or "redfish")}
}
\value{
Returns a tibble containing mean, sd, n, lwr, and upr values by year.
        Columns lwr and upr represent lower and upper 95 percent confidence intervals.
}
\description{
Function for calculating mean catch by year for species in the biomass data
}
