% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmb.coordinates.R
\name{gmb.coordinates}
\alias{gmb.coordinates}
\title{Convert genomic coordinates}
\usage{
gmb.coordinates(chr, pos, chrlen = mouse.chrlen)
}
\arguments{
\item{chr}{A character vector of chromosomes}

\item{pos}{A numeric vector of positions (b or Mb)}

\item{chrlen}{A vector with chromosome lengths (as in \code{org.Mm.egCHRLENGTHS})}
}
\value{
A numeric vector with cummulative (gmb) genomic coordinates.
}
\description{
Convert genomic coordinates from (chr, mb) representation to
cummulative (gmb) representation.
}
\details{
A cummulative (gmb) form could be useful for example as
x-coordinate of a genomic plot.
}
\examples{
gmb <- gmb.coordinates(Tmem68$annotation$chr, Tmem68$annotation$pos)
}
\seealso{
\code{\link{plot.mediation}}
}

