% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmb.coordinates.R
\name{gmb.coordinates}
\alias{gmb.coordinates}
\title{For chromosome and position, calculates genomic coordinate}
\usage{
gmb.coordinates(chr, pos, chrlen = mouse.chrlen)
}
\arguments{
\item{chr}{chromosome}

\item{pos}{position in bases or Mb}

\item{chrlen}{length of chromosomes as in \code{org.Mm.egCHRLENGTHS}}
}
\description{
For plotting, convert (chr, mb) coordinates to (gmb) coordinated
}

