% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identify.mediation.R
\name{identify.mediation}
\alias{identify.mediation}
\title{Identify Dot in Mediation Plot}
\usage{
\method{identify}{mediation}(medresults, label.col = "Associated.Gene.Name")
}
\arguments{
\item{medresults}{mediation object}

\item{label.col}{column of medresults to be plotted}
}
\description{
Identify Dot in Mediation Plot
}

