% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identify.mediation.R
\name{identify.mediation}
\alias{identify.mediation}
\title{Identify Dot in Mediation Plot}
\usage{
\method{identify}{mediation}(medresults, label.col = "Associated.Gene.Name")
}
\arguments{
\item{medresults}{mediation object}

\item{label.col}{column of medresults to be plotted}
}
\description{
Enables user to click on a plot and get a correcsponting
line of the annotation data frame.
}
\examples{
data(Tmem68)
med <- mediation.scan(target=Tmem68$target,
                      mediator=Tmem68$mediator,
                      annotation=Tmem68$annotation,
                      covar=Tmem68$covar,
                      qtl.geno=Tmem68$qtl.geno,
                      method="double-lod-diff")
plot(med, main="double-lod-diff")
identify(med)
}

