% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.mediation.R
\name{identify.mediation}
\alias{identify.mediation}
\title{Identify dot in mediation plot}
\usage{
\method{identify}{mediation}(med, label.col = "symbol")
}
\arguments{
\item{med}{A mediation object}

\item{label.col}{A column name of \code{med} to be used to be plotted}
}
\value{
A row from \code{med} corresponding to the point nearest to the click.
}
\description{
Enable user to identify a dot on a plot produced by \code{\link{plot.mediation}}.
}
\examples{
data(Tmem68)
med <- mediation.scan(target = Tmem68$target,
                      mediator = Tmem68$mediator,
                      annotation = Tmem68$annotation,
                      covar = Tmem68$covar,
                      qtl.geno = Tmem68$qtl.geno)
plot(med)
identify(med)
}
\seealso{
\code{\link{plot.mediation}}
}
