% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplot.R
\name{kplot}
\alias{kplot}
\title{Interactive plot of conditioned LOD scores}
\usage{
kplot(med, symbol.col = "symbol", chrlen = mouse.chrlen, ...)
}
\arguments{
\item{med}{A mediation object}

\item{symbol.col}{A column of \code{med} with gene names}

\item{chrlen}{A vector with chromosome lengths (as in \code{org.Mm.egCHRLENGTHS})}
}
\value{
Htmlwidget \code{iplot}
}
\description{
Plot LOD statistics calculated by \code{mediation.scan} against
genomic positions. Display a gene name when hovering a mouse over.
}
\examples{
data(Tmem68)
med <- mediation.scan(target = Tmem68$target,
                      mediator = Tmem68$mediator,
                      annotation = Tmem68$annotation,
                      covar = Tmem68$covar,
                      qtl.geno = Tmem68$qtl.geno)

theme_set(theme_bw())
kplot(med, symbol.col="symbol")

# save the plot as HTML page
gp <- kplot(med, symbol.col="symbol") 
pl <- ggplotly(gp, tooltip="text")
pl
tmpdir <- tempdir()
htmlwidgets::saveWidget(pl, paste0(tmpdir, "/Tmem68.html"),
                        selfcontained=TRUE)
                        
}
\seealso{
\code{\link{plot.mediation}}, \code{\link{mediation.scan}}
}
