% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mediation.scan.R
\name{mediation.scan}
\alias{mediation.scan}
\title{Mediation Scan}
\usage{
mediation.scan(target, mediator, annotation, qtl.geno, covar = NULL,
  method = c("ignore", "lod-diff", "double-lod-diff", "lod-ratio"),
  verbose = TRUE)
}
\arguments{
\item{target}{numeric vector with gene/protein expression}

\item{mediator}{matrix, each column is one gene/protein's expression}

\item{annotation}{data.frame with mediator annotation, must include columns "chr" and "pos"}

\item{qtl.geno}{matrix, haplotype probabilities at QTL we try to mediate}

\item{covar}{additive covariates}

\item{method}{select a method to handle missing cases}

\item{verbose}{if TRUE, display information about the progress}
}
\description{
For a given QTL haplotype probabilities `Q` and target `T`, the function tries to add all
possible mediators `M` and calculates LOD statistic. The low LOD value indicates `Q` and
`T` are conditionally independent given `M`, i.e. M is a mediator of causal relationship
for `Q` to T.
}
\examples{
data(Tmem68)
med <- mediation.scan(target=Tmem68$target,
                      mediator=Tmem68$mediator,
                      annotation=Tmem68$annotation,
                      covar=Tmem68$covar,
                      qtl.geno=Tmem68$qtl.geno,
                      method="double-lod-diff")
plot(med, main="double-lod-diff")
}

