% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.mediation.R
\name{plot.mediation}
\alias{plot.mediation}
\title{Plots conditioned LOD scores against genome position}
\usage{
\method{plot}{mediation}(medresults, col = "firebrick4",
  chrlen = mouse.chrlen, verbose = FALSE, ...)
}
\arguments{
\item{medresults}{mediation object}

\item{chrlen}{length of chromosomes as in \code{org.Mm.egCHRLENGTHS}}

\item{verbose}{if TRUE then prints "Finished!" when done}
}
\description{
For a given QTL haplotype probabilities `Q` and target `T`, the function tries to add all
possible mediators `M` and calculates LOD statistic. The low LOD value indicates `Q` and
`T` are conditionally independent given `M`, i.e. M is a mediator of causal relationship
for `Q` to T.
}
\examples{
data(Tmem68)
med <- mediation.scan(target=Tmem68$target,
                      mediator=Tmem68$mediator,
                      annotation=Tmem68$annotation,
                      covar=Tmem68$covar,
                      qtl.geno=Tmem68$qtl.geno,
                      method="double-lod-diff")
plot(med, main="double-lod-diff")
}
\author{
Steven C. Munger
}

