% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.mediation.R
\name{plot.mediation}
\alias{plot.mediation}
\title{Plots conditioned LOD scores against genome position}
\usage{
\method{plot}{mediation}(medresults, chrlen = mouse.chrlen, ...)
}
\arguments{
\item{medresults}{}

\item{chrlen}{length of chromosomes as in \code{org.Mm.egCHRLENGTHS}}
}
\description{
For a given QTL haplotype probabilities `Q` and target `T`, the function tries to add all
possible mediators `M` and calculates LOD statistic. The low LOD value indicates `Q` and
`T` are conditionally independent given `M`, i.e. M is a mediator of causal relationship
for `Q` to T.
}

