% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mediation.R
\name{plot.mediation}
\alias{plot.mediation}
\title{Plots conditioned LOD scores against genomic positions}
\usage{
\method{plot}{mediation}(med, col = "firebrick4", chrlen = mouse.chrlen,
  ...)
}
\arguments{
\item{med}{A mediation object}

\item{col}{A color of points}

\item{chrlen}{A vector with chromosome lengths (as in \code{org.Mm.egCHRLENGTHS})}
}
\description{
Plot LOD statistics calculated by \code{mediation.scan} against genomic positions.
}
\examples{
data(Tmem68)
med <- mediation.scan(target=Tmem68$target,
                      mediator=Tmem68$mediator,
                      annotation=Tmem68$annotation,
                      covar=Tmem68$covar,
                      qtl.geno=Tmem68$qtl.geno)
plot(med)
}
\seealso{
\code{\link{identify.mediation}}, \code{\link{kplot}}
}
